+++ mountpoint-s3/examples/fs_benchmark.rs
@@ -3,9 +3,8 @@ use fuser::{BackgroundSession, MountOption, Session};
 use mountpoint_s3::fuse::S3FuseFilesystem;
 use mountpoint_s3::prefetch::default_prefetch;
 use mountpoint_s3::{S3Filesystem, S3FilesystemConfig};
-use mountpoint_s3_client::config::{EndpointConfig, S3ClientConfig};
+use mountpoint_s3_client::config::{EndpointConfig, RustLogAdapter, S3ClientConfig};
 use mountpoint_s3_client::S3CrtClient;
-use mountpoint_s3_crt::common::rust_log_adapter::RustLogAdapter;
 use std::{
     fs::{File, OpenOptions},
     io::{self, BufRead, BufReader},