+++ mountpoint-s3/src/data_cache/disk_data_cache.rs
@@ -9,7 +9,7 @@ use std::time::Instant;
 use async_trait::async_trait;
 use bytes::Bytes;
 use linked_hash_map::LinkedHashMap;
-use mountpoint_s3_crt::checksums::crc32c::{self, Crc32c};
+use mountpoint_s3_client::checksums::crc32c::{self, Crc32c};
 use serde::{Deserialize, Serialize};
 use sha2::{Digest, Sha256};
 use thiserror::Error;