+++ mountpoint-s3-client/CHANGELOG.md
@@ -1,11 +1,15 @@
 ## Unreleased
 
+## v0.13.0 (February 3, 2025)
+
 ### Other changes
 
-* Add support for CRC64-NVME checksum algorithm. ([#1235](https://github.com/awslabs/mountpoint-s3/pull/1235))
+* Add support for CRC64-NVME checksum algorithm.
+  ([#1235](https://github.com/awslabs/mountpoint-s3/pull/1235))
 * Add support for overriding the number of threads used by the Event Loop.
   ([#1240](https://github.com/awslabs/mountpoint-s3/pull/1240))
-* The ECS credentials provider now performs retries in the event of some failures. ([awslabs/aws-c-auth#259](https://github.com/awslabs/aws-c-auth/pull/259))
+* The ECS credentials provider now performs retries in the event of some failures.
+  ([awslabs/aws-c-auth#259](https://github.com/awslabs/aws-c-auth/pull/259))
 * Export missing types in public API which previously required a direct dependency on `mountpoint-s3-crt`.
   ([#1248](https://github.com/awslabs/mountpoint-s3/pull/1248))